CARBON-15477: - Added utils package along with related classes - Temporarily commented code in profile gen and repo gen. (which will be fixed in a different task)
