+++ carbon-p2-plugin/src/main/java/org/wso2/maven/p2/CatFeature.java
@@ -21,7 +21,6 @@ import java.util.regex.Pattern;
 
 import org.apache.maven.plugin.MojoExecutionException;
 import org.apache.maven.project.MavenProject;
-import org.wso2.maven.p2.generate.feature.Bundle;
 
 public class CatFeature {
     
@@ -64,7 +63,8 @@ public class CatFeature {
     	if (!versionReplaced) {
     		replaceProjectKeysInVersion(project);
     	}
-        return Bundle.getOSGIVersion(version);
+        //return Bundle.getOSGIVersion(version);
+        return "";
     }
 
     public void setVersion(String version) {
