+++ carbon-p2-plugin/src/main/java/org/wso2/maven/p2/Category.java
@@ -65,17 +65,17 @@ public class Category {
 		return features;
 	}
 	
-    public ArrayList<CatFeature> getProcessedFeatures(MavenProject project, ArtifactFactory artifactFactory, List remoteRepositories, ArtifactRepository localRepository, ArtifactResolver resolver) throws MojoExecutionException{
+    public ArrayList<CatFeature> getProcessedFeatures(MavenProject project) throws MojoExecutionException{
         if (processedFeatures != null)
             return processedFeatures;
         if (features == null || features.size() == 0) return null;
         processedFeatures = new ArrayList<CatFeature>();
         Iterator<CatFeature> iter = features.iterator();
         while (iter.hasNext()) {
-            CatFeature f = (CatFeature)iter.next();       
+            CatFeature f = (CatFeature)iter.next();
             processedFeatures.add(f);
-            f.replaceProjectKeysInVersion(project); 
-        }        
+            f.replaceProjectKeysInVersion(project);
+        }
         return processedFeatures;
     }
