+++ carbon-p2-plugin/src/main/java/org/wso2/maven/p2/MaterializeProductMojo.java
@@ -23,8 +23,8 @@ import org.apache.maven.plugin.MojoFailureException;
 import org.apache.maven.project.MavenProject;
 import org.eclipse.tycho.model.ProductConfiguration;
 import org.eclipse.tycho.p2.facade.internal.P2ApplicationLauncher;
-import org.wso2.maven.p2.generate.utils.FileManagementUtil;
-import org.wso2.maven.p2.generate.utils.P2Constants;
+import org.wso2.maven.p2.utils.FileManagementUtil;
+import org.wso2.maven.p2.utils.P2Constants;
 
 import java.io.File;
 import java.io.FileOutputStream;
