+++ carbon-p2-plugin/src/main/java/org/wso2/maven/p2/ProfileGenMojo.java
@@ -30,10 +30,10 @@ import org.apache.maven.project.MavenProject;
 import org.apache.maven.project.MavenProjectHelper;
 import org.codehaus.plexus.util.FileUtils;
 import org.eclipse.tycho.p2.facade.internal.P2ApplicationLauncher;
-import org.wso2.maven.p2.generate.utils.FileManagementUtil;
-import org.wso2.maven.p2.generate.utils.MavenUtils;
-import org.wso2.maven.p2.generate.utils.P2Constants;
-import org.wso2.maven.p2.generate.utils.P2Utils;
+import org.wso2.maven.p2.utils.MavenUtils;
+import org.wso2.maven.p2.utils.FileManagementUtil;
+import org.wso2.maven.p2.utils.P2Constants;
+import org.wso2.maven.p2.utils.P2Utils;
 
 /**
  * Write environment information for the current build to file.
